# openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Nexus REST API",
    "description": "Explore resources and get their data.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/v1/jobs/{jobId}": {
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancels the specified job.",
        "operationId": "Jobs_CancelJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}/status": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Gets the status of the specified job.",
        "operationId": "Jobs_GetJobStatus",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/export": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Creates a new export job.",
        "operationId": "Jobs_Export",
        "requestBody": {
          "x-name": "parameters",
          "description": "Export parameters.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportParameters"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/clear-cache": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Clears the aggregation data cache for the specified period of time.",
        "operationId": "Jobs_ClearCache",
        "parameters": [
          {
            "name": "catalogId",
            "in": "query",
            "required": true,
            "description": "The catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "begin",
            "in": "query",
            "required": true,
            "description": "Start date/time.",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-position": 2
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "description": "End date/time.",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Job": {
        "type": "object",
        "description": "Description of a job.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "The global unique identifier.",
            "format": "guid"
          },
          "type": {
            "type": "string",
            "description": "The job type."
          },
          "owner": {
            "type": "string",
            "description": "The owner of the job."
          },
          "parameters": {
            "description": "The job parameters.",
            "nullable": true
          }
        }
      },
      "JobStatus": {
        "type": "object",
        "description": "Describes the status of the job.",
        "additionalProperties": false,
        "properties": {
          "start": {
            "type": "string",
            "description": "The start date/time.",
            "format": "date-time"
          },
          "status": {
            "description": "The status.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              }
            ]
          },
          "progress": {
            "type": "number",
            "description": "The progress from 0 to 1.",
            "format": "double"
          },
          "exceptionMessage": {
            "type": "string",
            "description": "The nullable exception message.",
            "nullable": true
          },
          "result": {
            "description": "The nullable result.",
            "nullable": true
          }
        }
      },
      "TaskStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Created",
          "WaitingForActivation",
          "WaitingToRun",
          "Running",
          "WaitingForChildrenToComplete",
          "RanToCompletion",
          "Canceled",
          "Faulted"
        ],
        "enum": [
          "Created",
          "WaitingForActivation",
          "WaitingToRun",
          "Running",
          "WaitingForChildrenToComplete",
          "RanToCompletion",
          "Canceled",
          "Faulted"
        ]
      },
      "ExportParameters": {
        "type": "object",
        "description": "A structure for export parameters.",
        "additionalProperties": false,
        "properties": {
          "begin": {
            "type": "string",
            "description": "The start date/time.",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "description": "The end date/time.",
            "format": "date-time"
          },
          "filePeriod": {
            "type": "string",
            "description": "The file period.",
            "format": "duration"
          },
          "type": {
            "type": "string",
            "description": "The writer type. If null, data will be read (and possibly cached) but not returned. This is useful for data pre-aggregation.",
            "nullable": true
          },
          "resourcePaths": {
            "type": "array",
            "description": "The resource paths to export.",
            "items": {
              "type": "string"
            }
          },
          "configuration": {
            "type": "object",
            "description": "The configuration.",
            "nullable": true,
            "additionalProperties": {}
          }
        }
      },
...
    }
  }
}